+++ mopidy_audioaddict/actor.py
@@ -91,12 +91,6 @@ class AudioAddictLibrary(backend.LibraryProvider):
 
         return [Track(uri=ref.uri, name=ref.name)]
 
-    def find_exact(self, query=None, uris=None):
-        return None
-
-    def search(self, query=None, uris=None):
-        return None
-
 
 class AudioAddictPlayback(backend.PlaybackProvider):
     def change_track(self, track):