+++ CHANGES.rst
@@ -6,6 +6,8 @@ v0.3.0 (UNRELEASED)
 
 - Require Mopidy >= 1.0
 
+- Updated to work with backend API changes in Mopidy 1.0
+
 
 v0.2.7 (2014-11-03)
 -------------------